Use current output configuration
Do not trigger re-enumerating output devices when changing or querying
the backlight.

Re-enumerating output devices may stall Xorg at least on Intel GPUs when
EDID is unresponsive or contains bogus data (due to retries). When
working with the backlight it is safe to assume that a monitor capable
of setting its backlight is currently connected.

Signed-off-by: Nils Schneider <nils@nilsschneider.net>
Reviewed-by: Hans de Goede <hdegoede@redhat.com>
Signed-off-by: Hans de Goede <hdegoede@redhat.com>